+++ CHANGELOG.md
@@ -27,6 +27,7 @@
  - New component: WebServer. A more general way of server content through http. Can replace
    WebWidget, so WebWidget will be deprecated
  - New component: WebSocket. A websocket client.
+ - New IRacing event: IRacingTime - sent every second. Contains session time and session time remaining
 
 ## [0.10.0](https://github.com/dennis/slipstream/releases/tag/v0.10.0) (2021-10-07)
 [Full Changelog](https://github.com/dennis/slipstream/compare/v0.9.0...v0.10.0)